+++ CHANGELOG.md
@@ -19,3 +19,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 - Moved the main block parser functionality so that it is now available at `Blocky::$plugin->parseBlocks()`
   - This is consumed by the new `{% blocks %}` Twig tag and `{% craft.blocky.blockparser() %}`
+
+## 0.1.1 - 2020-07-05
+### Changed
+- Typo in class name `InvalicBlockException`.
+- The class Blocky is not correctly importing the exception classes.
+
+### Removed
+- Translation files, as these only contained exception translations.
+- All instances of `Craft::t()` in exceptions.